Release candidate v1.1.0

+++ bootstrap.ps1
@@ -1,4 +1,5 @@
 function Add-EnvPath {
+    # Get parameters
     param(
         [Parameter(Mandatory=$true)]
         [string] $Path,
@@ -6,7 +7,16 @@ function Add-EnvPath {
         [ValidateSet('Machine', 'User', 'Session')]
         [string] $Container = 'Session'
     )
+    
+    # Do it for global Path (like SETX but no restriction over 1024 characters)
+    if(!($env:Path -like "*$Path*")) {
+        "[INFO] Adding $Path to system path!"
+        reg add "HKLM\SYSTEM\CurrentControlSet\Control\Session Manager\Environment" /f /v Path /t REG_SZ /d "$env:Path;$Path;"
+    } else {
+        "[INFO] $Path is already in system path."
+    }
 
+    # Do it for current $env:Path
     if ($Container -ne 'Session') {
         $containerMapping = @{
             Machine = [EnvironmentVariableTarget]::Machine
@@ -27,6 +37,9 @@ function Add-EnvPath {
         $env:Path = $envPaths -join ';'
     }
 }
+
+# back up the current environment variables
+"$env:Path" | Out-File "Path-Backup-$(Get-Date -format yyyyMMddHHmmss).txt"
 
 ########################
 #                      #
@@ -70,11 +83,29 @@ If ($lastexitcode -notmatch 0) {
         "[ERROR] There were problems when installing git."
         Exit
     }
-    SETX PATH "C:\Program Files\Git\bin;$env:Path" /m
-    Add-EnvPath "C:\Program Files\Git\bin;"
+    Add-EnvPath "C:\Program Files\Git\bin"
 } Else {
     "[INFO] Git found in current system. Not going to install it."
 }
+
+
+# Install OBS Studio
+If (Test-Path "C:\Program Files (x86)\obs-studio\bin\32bit\obs32.exe") {
+    "[INFO] OBS had been installed."
+} Else {
+    "[INFO] Downloading OBS ..."
+    pushd thirdParty
+    If (Test-Path "OBS-Studio-18.0.1-Full.zip") {
+        Remove-Item "OBS-Studio-18.0.1-Full.zip"
+    }
+    CMD /C curl -kLO https://github.com/jp9000/obs-studio/releases/download/18.0.1/OBS-Studio-18.0.1-Full.zip
+    "[INFO] Installing OBS ..."
+    Expand-Archive -LiteralPath OBS-Studio-18.0.1-Full.zip -DestinationPath "C:\Program Files (x86)\obs-studio"
+    popd
+    $FIRST_OBS = $TRUE
+    "[INFO] OBS is installed."
+}
+
 
 # installation of Hasal prerequisite
 $tmp = & CMD /C java -version >$null 2>&1
@@ -85,8 +116,7 @@ If ($lastexitcode -notmatch 0) {
         "[ERROR] There were problems when installing jre8."
         Exit
     }
-    SETX PATH "C:\Program Files\Java\jre1.8.0_121\bin;$env:Path" /m
-    Add-EnvPath "C:\Program Files\Java\jre1.8.0_121\bin;"
+    Add-EnvPath "C:\Program Files\Java\jre1.8.0_121\bin"
 } Else {
     "[INFO] Java found in current system. Not going to install it."
 }
@@ -100,8 +130,7 @@ If ($lastexitcode -notmatch 0) {
         "[ERROR] There were problems when installing ffmpeg."
         Exit
     }
-    SETX PATH "C:\ProgramData\chocolatey\lib\ffmpeg\tools\ffmpeg-3.2.2-win32-static\bin\;$env:Path" /m
-    Add-EnvPath "C:\ProgramData\chocolatey\lib\ffmpeg\tools\ffmpeg-3.2.2-win32-static\bin\;"
+    Add-EnvPath "C:\ProgramData\chocolatey\lib\ffmpeg\tools\ffmpeg-3.2.2-win32-static\bin\"
 } Else {
     "[INFO] FFMPEG found in current system. Not going to install it."
 }
@@ -114,8 +143,8 @@ If ($lastexitcode -notmatch 0) {
         "[ERROR] There were problems when installing miniconda."
         Exit
     }
-    SETX PATH "C:\Program Files\Miniconda2\Scripts\;C:\Program Files\Miniconda2\;$env:Path" /m
-    Add-EnvPath "C:\Program Files\Miniconda2\Scripts\;C:\Program Files\Miniconda2\;"
+    Add-EnvPath "C:\Program Files\Miniconda2\Scripts\"
+    Add-EnvPath "C:\Program Files\Miniconda2\"
 } Else {
     "[INFO] MiniConda found in current system. Not going to install it."
 }
@@ -129,7 +158,10 @@ If ($lastexitcode -notmatch 0) {
     Exit
 }
 "[INFO] Removing VCForPython27.msi"
-Remove-Item VCForPython27.msi
+If (Test-Path "VCForPython27.msi") {
+    Remove-Item VCForPython27.msi
+}
+
 
 # installation of browsers
 "[INFO] Installing Chrome"
@@ -138,9 +170,12 @@ If ($lastexitcode -notmatch 0) {
     "[ERROR] There were problems when installing Chrome"
     Exit
 }
-SETX PATH "C:\Program Files\Google\Chrome\Application\;C:\Program Files (x86)\Google\Chrome\Application\;$env:Path" /m
-Add-EnvPath "C:\Program Files\Google\Chrome\Application\;"
-Add-EnvPath "C:\Program Files (x86)\Google\Chrome\Application\;"
+
+If (Test-Path "C:\Program Files\Google\Chrome\Application\") {
+    Add-EnvPath "C:\Program Files\Google\Chrome\Application\"
+} ElseIf (Test-Path "C:\Program Files (x86)\Google\Chrome\Application\") {
+    Add-EnvPath "C:\Program Files (x86)\Google\Chrome\Application\"
+}
 
 "[INFO] Installing Firefox"
 choco install firefox -y
@@ -148,9 +183,12 @@ If ($lastexitcode -notmatch 0) {
     "[ERROR] There were problems when installing Firefox"
     Exit
 }
-SETX PATH "C:\Program Files\Mozilla Firefox;C:\Program Files (x86)\Mozilla Firefox;$env:Path" /m
-Add-EnvPath "C:\Program Files\Mozilla Firefox;"
-Add-EnvPath "C:\Program Files (x86)\Mozilla Firefox;"
+
+If (Test-Path "C:\Program Files\Mozilla Firefox") {
+    Add-EnvPath "C:\Program Files\Mozilla Firefox"
+} ElseIf (Test-Path "C:\Program Files (x86)\Mozilla Firefox") {
+    Add-EnvPath "C:\Program Files (x86)\Mozilla Firefox"
+}
 
 # installation of sikuli
 "[INFO] Installing SikuliX 1.1.0"
@@ -169,7 +207,7 @@ If ($lastexitcode -notmatch 0) {
 "[INFO] Creating Miniconda virtualenv (It might take some time to finish.)"
 conda config --set always_yes yes --set changeps1 no
 conda install psutil
-conda create -q -n env-python python=2.7 numpy scipy nose pywin32 pip
+conda create -q -n env-python python=2.7 numpy scipy nose pywin32 pip matplotlib==1.5.3 runipy==0.1.5
 
 Copy-Item C:\ProgramData\chocolatey\lib\sikulix\content\* .\thirdParty\
 Copy-Item .\scripts\runsikuli* .\thirdParty\
@@ -189,3 +227,28 @@ If (Test-Path C:\Miniconda2\) {
     CMD /C "C:\Program Files\Miniconda2\envs\env-python\python" setup.py install
     CMD /C "C:\Program Files\Miniconda2\envs\env-python\python" scripts\cv2_checker.py
 }
+
+
+########################
+#                      #
+#   User Interactions  #
+#                      #
+########################
+
+# OBS License Agreement
+IF ($FIRST_OBS) {
+    If (Test-Path "C:\Program Files (x86)\obs-studio\bin\32bit\obs32.exe") {
+        pushd "C:\Program Files (x86)\obs-studio\bin\32bit\"
+        "[INFO] Launching OBS for License Agreement."
+        ""
+        "[INFO] If there is error message about missing MSVCP120.dll,"
+        "[INFO] please install Visual C++ Redistributable Packages from https://www.microsoft.com/en-us/download/details.aspx?id=40784"
+        "[INFO] ** Please CLOSE OBS after accepting License Agreement **"
+        CMD /C obs32.exe
+        popd
+    } ELSE {
+        "[WARN] Can not find OBS binary file."
+    }
+}
+
+"[INFO] You will need to restart the computer so that the PATH will be effective"
